+++ example/lib/posts/view_model/posts_effect.dart
@@ -1,6 +1,0 @@
-import 'package:mvi/mvi.dart';
-
-sealed class PostsEffect extends BaseEffect {
-  @override
-  String toString() => 'PostsEffect';
-}