+++ examples/counter/lib/counter/counter_page.dart
@@ -1,0 +1,65 @@
+import 'package:counter/counter/view_model/counter_view_model.dart';
+import 'package:flutter/material.dart';
+import 'package:mvi/mvi.dart';
+
+class CounterPage extends StatefulWidget {
+  const CounterPage({super.key});
+
+  @override
+  State<CounterPage> createState() => _CounterPageState();
+}
+
+class _CounterPageState extends State<CounterPage>
+    with
+        SimpleViewModelMixin<
+          CounterPage,
+          CounterState,
+          CounterEvent,
+          CounterViewModel
+        > {
+  @override
+  CounterViewModel createViewModel() => CounterViewModel();
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+        title: const Text('Counter'),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            const Text('You have pushed the button this many times:'),
+            ValueListenableBuilder(
+              valueListenable: viewModel.state,
+              builder: (context, state, child) {
+                return Text(
+                  '${state.count}',
+                  style: Theme.of(context).textTheme.headlineMedium,
+                );
+              },
+            ),
+          ],
+        ),
+      ),
+      floatingActionButton: Column(
+        mainAxisAlignment: MainAxisAlignment.end,
+        children: [
+          FloatingActionButton(
+            onPressed: () => addEvent(Increment()),
+            tooltip: 'Increment',
+            child: const Icon(Icons.add),
+          ),
+          const SizedBox(height: 8),
+          FloatingActionButton(
+            onPressed: () => addEvent(Decrement()),
+            tooltip: 'Decrement',
+            child: const Icon(Icons.remove),
+          ),
+        ],
+      ),
+    );
+  }
+}