+++ examples/counter/lib/counter/view_model/counter_event.dart
@@ -1,0 +1,13 @@
+import 'package:mvi/mvi.dart';
+
+sealed class CounterEvent extends BaseEvent {}
+
+final class Increment extends CounterEvent {
+  @override
+  String toString() => 'Increment';
+}
+
+final class Decrement extends CounterEvent {
+  @override
+  String toString() => 'Decrement';
+}